+++ src/test/kotlin/net/pwall/json/schema/codegen/test/kotlin/AdditionalPropertiesDeserializationTest.kt
@@ -25,14 +25,16 @@
 
 package net.pwall.json.schema.codegen.test.kotlin
 
-import io.kjson.JSONKotlinException
-import io.kjson.parseJSON
-import java.time.LocalDate
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
 import kotlin.test.assertIs
 import kotlin.test.assertNull
 import kotlin.test.expect
+
+import java.time.LocalDate
+
+import io.kjson.JSONKotlinException
+import io.kjson.parseJSON
 
 class AdditionalPropertiesDeserializationTest {
 
@@ -94,31 +96,31 @@ class AdditionalPropertiesDeserializationTest {
 
     @Test fun `should deserialize TestApTrueExtra`() {
         val json1 = """{"extra":"content"}"""
-        val result1 = json1.parseJSON<TestApTrueExtra>()
+        val result1 = json1.parseJSON<TestApTrueExtraValid>()
         expect("content") { result1.extra }
         assertNull(result1["whatever"])
 
         val json2 = """{"extra":"content","anything":"another"}"""
-        val result2 = json2.parseJSON<TestApTrueExtra>()
+        val result2 = json2.parseJSON<TestApTrueExtraValid>()
         expect("content") { result2.extra }
         expect("another") { result2["anything"] }
         assertNull(result2["whatever"])
 
         assertFailsWith<JSONKotlinException> {
-            """{"anything":"another"}""".parseJSON<TestApTrueExtra>()
+            """{"anything":"another"}""".parseJSON<TestApTrueExtraValid>()
         }.cause.let {
             assertIs<IllegalArgumentException>(it)
             expect("required property missing - extra") { it.message }
         }
 
         assertFailsWith<JSONKotlinException> {
-            """{"extra":123}""".parseJSON<TestApTrueExtra>()
+            """{"extra":123}""".parseJSON<TestApTrueExtraValid>()
         }.let {
-            expect("Can't deserialize 123 as String at /extra") { it.message }
+            expect("Incorrect type, expected string but was 123, at /extra") { it.message }
         }
 
         assertFailsWith<JSONKotlinException> {
-            """{"extra":""}""".parseJSON<TestApTrueExtra>()
+            """{"extra":""}""".parseJSON<TestApTrueExtraValid>()
         }.cause.let {
             assertIs<IllegalArgumentException>(it)
             expect("extra length < minimum 1 - 0") { it.message }
@@ -146,11 +148,11 @@ class AdditionalPropertiesDeserializationTest {
         assertFailsWith<JSONKotlinException> {
             """{"extra":123}""".parseJSON<TestApTrueExtraDefault>()
         }.let {
-            expect("Can't deserialize 123 as String at /extra") { it.message }
+            expect("Incorrect type, expected string but was 123, at /extra") { it.message }
         }
 
         assertFailsWith<JSONKotlinException> {
-            """{"extra":""}""".parseJSON<TestApTrueExtra>()
+            """{"extra":""}""".parseJSON<TestApTrueExtraValid>()
         }.cause.let {
             assertIs<IllegalArgumentException>(it)
             expect("extra length < minimum 1 - 0") { it.message }
@@ -189,7 +191,7 @@ class AdditionalPropertiesDeserializationTest {
         assertFailsWith<JSONKotlinException> {
             """{"extra":123}""".parseJSON<TestAdditionalPropertiesSchemaExtra>()
         }.let {
-            expect("Can't deserialize 123 as String at /extra") { it.message }
+            expect("Incorrect type, expected string but was 123, at /extra") { it.message }
         }
 
         assertFailsWith<JSONKotlinException> {